# api/comment.go
package api

import (
	"github.com/gin-gonic/gin"
	"github.com/intyouss/Traceability/models"
	"github.com/intyouss/Traceability/service"
	"github.com/intyouss/Traceability/service/dto"
	"github.com/jinzhu/copier"
)

const (
	ErrCodeGetCommentList = iota + 20001
	ErrCodeAddComment
	ErrCodeDeleteComment
)

type CommentApi struct {
	BaseApi
	UserApi
	Service *service.CommentService
}

func NewCommentApi() CommentApi {
	return CommentApi{
		BaseApi: NewBaseApi(),
		UserApi: NewUserApi(),
		Service: service.NewCommentService(),
	}
}

// GetCommentList 获取评论列表
// @Summary 获取评论列表
// @Description 获取评论列表
// @Param page formData int false "页码"
// @Param limit formData int false "每页数量"
// @Param video_id formData int true "视频id"
// @Success 200 {string} Response
// @Failure 400 {string} Response
// @Router /api/v1/public/comment/list [post]
func (c CommentApi) GetCommentList(ctx *gin.Context) {
	var cListDto dto.CommentListDTO
	if err := c.BuildRequest(BuildRequestOption{Ctx: ctx, DTO: &cListDto}).GetError(); err != nil {
		c.Fail(&Response{Msg: err.Error()})
		return
	}

	commentsDao, total, err := c.Service.GetCommentList(&cListDto)
	if err != nil {
		c.ServerError(&Response{Code: ErrCodeGetCommentList, Msg: err.Error()})
		return
	}

	var commentUserMap map[uint]*models.User
	for _, comment := range commentsDao {
		commentUserMap[comment.UserId] = nil
	}

	var userIds []uint
	for userId := range commentUserMap {
		userIds = append(userIds, userId)
	}

	users, err := c.UserApi.Service.GetUserListByIds(userIds)
	if err != nil {
		c.ServerError(&Response{Code: ErrCodeGetCommentList, Msg: err.Error()})
		return
	}

	for _, user := range users {
		commentUserMap[user.ID] = user
	}

	var comments = make([]dto.Comment, len(commentsDao))
	for i, comment := range commentsDao {
		comments[i].ID = comment.ID
		comments[i].Content = comment.Content
		comments[i].CreateAt = comment.CreatedAt.Format("2006-01-02 15:04:05")
		var user = new(dto.User)
		_ = copier.Copy(user, commentUserMap[comment.UserId])
		comments[i].User = user
	}

	c.Success(&Response{
		Data:  comments,
		Total: total,
	})
}

// AddComment 添加评论
// @Summary 添加评论
// @Description 添加评论
// @Param video_id formData int true "视频id"
// @Param content formData string false "评论内容"
// @Success 200 {string} Response
// @Failure 400 {string} Response
// @Router /api/v1/comment/add [post]
func (c CommentApi) AddComment(ctx *gin.Context) {
	var cAddDto dto.AddCommentDTO
	if err := c.BuildRequest(BuildRequestOption{Ctx: ctx, DTO: &cAddDto}).GetError(); err != nil {
		c.Fail(&Response{Msg: err.Error()})
		return
	}
	commentDao, err := c.Service.AddComment(&cAddDto)
	if err != nil {
		c.ServerError(&Response{Code: ErrCodeAddComment, Msg: err.Error()})
		return
	}
	var IdDTO dto.CommonUserIDDTO
	IdDTO.ID = commentDao.UserId
	userDao, err := c.UserApi.Service.GetUserById(&IdDTO)
	if err != nil {
		c.ServerError(&Response{Code: ErrCodeAddComment, Msg: err.Error()})
		return
	}

	var comment = new(dto.Comment)
	_ = copier.Copy(comment, commentDao)
	var user = new(dto.User)
	_ = copier.Copy(user, userDao)
	comment.User = user

	c.Success(&Response{
		Data: comment,
	})
}

// DeleteComment 删除评论
// @Summary 删除评论
// @Description 删除评论
// @Param comment_id formData int true "评论id"
// @Success 200 {string} Response
// @Failure 400 {string} Response
// @Router /api/v1/comment/delete [post]
func (c CommentApi) DeleteComment(ctx *gin.Context) {
	var cDeleteDto dto.DeleteCommentDTO
	if err := c.BuildRequest(BuildRequestOption{Ctx: ctx, DTO: &cDeleteDto}).GetError(); err != nil {
		c.Fail(&Response{Msg: err.Error()})
		return
	}
	err := c.Service.DeleteCommentById(&cDeleteDto)
	if err != nil {
		c.ServerError(&Response{Code: ErrCodeDeleteComment, Msg: err.Error()})
		return
	}
	c.Success(&Response{})
}