# docs/swagger.yaml
info:
  contact: {}
  description: A web-based short video platform
  title: Traceability
  version: 0.0.1
paths:
  /api/v1/comment/add:
    post:
      description: 添加评论
      parameters:
      - description: 视频id
        in: formData
        name: video_id
        required: true
        type: integer
      - description: 评论内容
        in: formData
        name: content
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 添加评论
  /api/v1/comment/delete:
    post:
      description: 删除评论
      parameters:
      - description: 评论id
        in: formData
        name: comment_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 删除评论
  /api/v1/public/comment/list:
    post:
      description: 获取评论列表
      parameters:
      - description: 页码
        in: formData
        name: page
        type: integer
      - description: 每页数量
        in: formData
        name: limit
        type: integer
      - description: 视频id
        in: formData
        name: video_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 获取评论列表
  /api/v1/public/user/login:
    post:
      description: 用户登录
      parameters:
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
        "401":
          description: Unauthorized
          schema:
            type: string
      summary: 用户登录
  /api/v1/public/user/register:
    post:
      description: 用户注册
      parameters:
      - description: 用户名
        in: formData
        name: username
        required: true
        type: string
      - description: 密码
        in: formData
        name: password
        required: true
        type: string
      - description: 邮箱
        in: formData
        name: email
        type: string
      - description: 手机号
        in: formData
        name: mobile
        type: string
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 用户注册
  /api/v1/public/video/feed:
    post:
      description: 获取视频feed流
      parameters:
      - description: 页码
        in: formData
        name: page
        type: integer
      - description: 每页数量
        in: formData
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 获取视频feed流
  /api/v1/user/:
    get:
      description: 获取用户信息
      parameters:
      - description: 用户id
        in: query
        name: user_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 获取用户信息
  /api/v1/user/delete:
    post:
      description: 删除用户
      parameters:
      - description: 用户id
        in: formData
        name: user_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 删除用户
  /api/v1/user/list:
    post:
      description: 获取用户列表
      parameters:
      - description: 页码
        in: formData
        name: page
        type: integer
      - description: 每页数量
        in: formData
        name: limit
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 获取用户列表
  /api/v1/user/update:
    post:
      description: 更新用户信息
      parameters:
      - description: 用户id
        in: formData
        name: id
        required: true
        type: integer
      - description: 用户名
        in: formData
        name: username
        type: string
      - description: 密码
        in: formData
        name: password
        type: string
      - description: 邮箱
        in: formData
        name: email
        type: string
      - description: 手机号
        in: formData
        name: mobile
        type: string
      - description: 头像
        in: formData
        name: avatar
        type: file
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 更新用户信息
  /api/v1/video/delete:
    delete:
      description: 删除视频
      parameters:
      - description: 用户id
        in: formData
        name: id
        required: true
        type: integer
      - description: 视频id
        in: formData
        name: video_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 删除视频
  /api/v1/video/list:
    get:
      description: 获取用户发布视频列表
      parameters:
      - description: 用户id
        in: query
        name: user_id
        required: true
        type: integer
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 获取用户发布视频列表
  /api/v1/video/publish:
    post:
      description: 发布视频
      parameters:
      - description: 用户id
        in: formData
        name: id
        required: true
        type: integer
      - description: 视频标题
        in: formData
        name: title
        required: true
        type: string
      - description: 视频文件
        in: formData
        name: data
        required: true
        type: file
      responses:
        "200":
          description: OK
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            type: string
      summary: 发布视频
swagger: "2.0"